+++ CHANGELOG.md
@@ -10,12 +10,13 @@ to [Common Changelog](https://common-changelog.org)
 
 ### Changed
 
-- Display Venn diagram as Edwards for 6 genomes, as numbers were missing in the classic representation. ([#132](https://github.com/metagenlab/zDB/pull/132)) (Niklaus Johner)
-- Add contig accession and range to genomic regions in plot_region. ([#131](https://github.com/metagenlab/zDB/pull/131)) (Niklaus Johner)
+- Display Venn diagram as Edwards for 6 genomes, as numbers were missing in the classic representation. ([#135](https://github.com/metagenlab/zDB/pull/135)) (Niklaus Johner)
+- Add contig accession and range to genomic regions in plot_region. ([#134](https://github.com/metagenlab/zDB/pull/134)) (Niklaus Johner)
 - Improve error handling for missing input genomes and reference databases. ([#130](https://github.com/metagenlab/zDB/pull/130)) (Niklaus Johner)
 
 ### Added
 
+- Add contig card to locusx view. ([#136](https://github.com/metagenlab/zDB/pull/136)) (Niklaus Johner)
 - Add option to highlight certain leaves in custom plots. ([#131](https://github.com/metagenlab/zDB/pull/131)) (Niklaus Johner)