+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to [Common Changelog](https://common-changelog.org)
 ### Fixed
 
 - Fix annotation pipeline for new AMRFinder version 4. ([#141](https://github.com/metagenlab/zDB/pull/141)) (Niklaus Johner)
+- Fix DB setup pipeline when using containers and an unmounted reference directory. ([#142](https://github.com/metagenlab/zDB/pull/142)) (Niklaus Johner)
 
 ### Changed